# docs/openapi.yaml
openapi: 3.0.3
info:
  title: АИС «Реестр обязательных требований» — HTTP API
  version: "1.0.0"
  description: |
    Спецификация фактического API веб-приложения (Next.js App Router, 46 маршрутов).
    Составлена по коду `src/app/api/**` (2026-07-02), см. docs/architecture (комплект ГОСТ 34).

    **Аутентификация:** JWT в httpOnly-cookie `auth_token` (выдаётся `/api/auth/login`, срок 7 суток).
    **Роли:** `admin` (супер-админ МНЭ) · `mne` (сотрудник МНЭ: всё как admin, кроме выгрузки каталога и создания учёток уровня МНЭ) · `moderator` (орган + поддерево организаций) ·
    `expert` (назначенные сферы ∩ органы). Скоупы возвращает `/api/auth/me`.
    **Ошибки:** единый формат `{ "error": string }`; 401 — не авторизован, 403 — нет прав/чужой орган,
    404 — не найдено, 429 — превышен лимит (голосование), 5xx — внутренняя/внешний сервис.

    Раздел «Легаси-экспертиза» обслуживает исторический MVP-контур (requirement_cards/голосования)
    и не развивается — см. документ 09 плана (К5).
servers:
  - url: https://reestr-web-production.up.railway.app
    description: Продуктив (Railway)
  - url: http://localhost:3000
    description: Локальная разработка
security:
  - cookieAuth: []
...
paths:
...
  /api/registry/review:
    post:
      tags: [Ревью и АРА]
      summary: Действие по карточке(ам) реестра
      description: |
        `confirm` требует `ara_deadline` (≤3 лет для законов, ≤2 — прочие).
        `edit` принимает новые `title`/`legal_text` (+ запись в registry_edits).
        `include` — только admin (включение в реестр: included_at, ara_status).
        Массовый режим — `ids[]`. Доступ: authority_code карточки ∈ скоуп пользователя.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                id: { type: integer, description: Одна карточка }
                ids: { type: array, items: { type: integer }, description: Массовое действие }
                action: { type: string, enum: [confirm, reject, edit, include] }
                comment: { type: string }
                ara_deadline: { type: string, format: date, description: Обязателен при confirm }
                title: { type: string, description: Для edit }
                legal_text: { type: string, description: Для edit }
      responses:
        "200": { description: Применено (счётчик обновлённых) }
        "400": { description: Нет ara_deadline при confirm / неверное действие }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ══════════════════ Подача НПА ══════════════════
...
  /api/votes:
    post:
      tags: [Легаси-экспертиза]
      summary: Голос v1 (requirementId)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                requirementId: { type: integer }
                vote: { type: string }
                comment: { type: string }
      responses: { "200": { description: OK } }
    put:
      tags: [Легаси-экспертиза]
      summary: Массовые голоса v1
      responses: { "200": { description: OK } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT (7 суток), выдаётся /api/auth/login
  responses:
...
    Forbidden:
      description: Недостаточно прав / чужой орган
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...